requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
celery==5.3.4
opencv-python==4.8.1.78
pillow==10.1.0
transformers==4.35.0
pypdf2==3.0.1
pymupdf==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
easyocr==1.7.1
langchain==0.0.340
chromadb==0.4.18
google-generativeai==0.3.0
reportlab==4.0.7
xlsxwriter==3.1.9
pygithub==2.1.1
gitpython==3.1.40
httpx==0.25.1
aiohttp==3.9.1
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
networkx==3.2.1
